Added `arithmetic_kernels`
